streams: fix matching remote tracks
We generate random UUIs for our local tracks, but collisions are
possible for remote tracks. The unicity of tracks IDs is guaranteed only
in the scope of a single PeerConnection.

Thus, use the PeerConnection internal ID when we go looking for a track.
For local tracks we'll use an ID of -1.

Fixes: https://github.com/react-native-webrtc/react-native-webrtc/issues/1347